[XUI-3193] Fix calling select on border div (#6017)
Initially reported via [Slack](https://xero.slack.com/archives/C565NP1A5/p1646778448015539)

When selecting the border of a text input cell, we would attempt to call `select()` on the `target`. This resulted in an error, as the border could not be selected.

This change means we always call the select method on the text input's ref, so that we are selecting the correct thing regardless of whether the border is clicked.