[XUI-2714] Lower specifity of min-width rule
The way we set `min-width` in our buttons worked; but it removed the
ability for consumers to set their own `min-width` by applying a class
name to `XUIButton` – the `:not` raised the specificity higher than the
rest of the button rules.

I realise that clobbering styles is not best practice, but since this is
a small considered change with a limited blast radius I think it’s okay.