[XUI-3116] Update date picker Storybook playground
I added some missing props and converted the example from class-based to
functional with hooks, because I’m silly like that. I didn’t add a knob
for `isDateDisabled`, since it takes a function and it’s hard to see the
effect in the playground anyway.

Since there was some overlap between the work here and the work done on
the date input playground, I also abstracted out a helper I wrote to
tidy things up a little.