[XUI-3117] Dropdown Storybook improvements
This updates knobs and groupings, and also tweaks the content to be an
‘example of good’. Because the stories file was really annoying to try
to work with with all the vis-reg stuff in the way, I moved all of it
into a helper file to try to make things more manageable. This resulted
in a few minor changes to the reference images.

You’ll notice `showHeader` in more conditionals than you may feel
appropriate. I used that as a proxy for ‘wider multiselect dropdown‘
instead of setting up a whole other example.